crawler/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
openai==0.28.1
schedule==1.2.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
python-dateutil==2.8.2
python-dotenv==1.0.0
aiohttp==3.9.1
pillow==10.1.0
apscheduler==3.10.4
